backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyotp==2.9.0
python-multipart==0.0.18
redis[hiredis]==5.2.1
celery[redis]==5.4.0
celery-redbeat==2.2.0
firebase-admin==6.6.0
boto3==1.35.81
slowapi==0.1.9
qrcode[pil]==8.0
sendgrid==6.11.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp==1.29.0
prometheus-client==0.21.1
sentry-sdk[fastapi]==2.19.2
structlog==24.4.0
python-json-logger==3.2.1
Pillow==11.0.0
reportlab==4.2.5
aiofiles==24.1.0
websockets==14.1
